Module nds_sys::video

Video related functions and constants. Mostly “borrowed” from libnds’s video.h

Structs

Flags

Constants

BG_GFX

BG_GFX_SUB

BG_PALETTE

Background palette (Main)

BG_PALETTE_SUB

Background palette (Sub)

MAP_BASE_MASK

MAP_BASE_OFFSET

REG_DISPCNT

Display control register (main). Used to choose the mode for the Main Engine

REG_DISPCNT_SUB

Display control register (sub). Used to choose the mode for the Sub Engine

TILES_BASE_MASK

TILES_BASE_OFFSET

VRAM_A

When VRAM A is in LCD mode, this points to the first pixel (top left most)

VRAM_A_CR

Control register for VRAM A

VRAM_B

When VRAM B is in LCD mode, this points to the first pixel (top left most)

VRAM_B_CR

Control register for VRAM B

VRAM_C

When VRAM C is in LCD mode, this points to the first pixel (top left most)

VRAM_C_CR

Control register for VRAM C

VRAM_D

When VRAM D is in LCD mode, this points to the first pixel (top left most)

VRAM_D_CR

Control register for VRAM D

VRAM_ENABLE

VRAM_E_CR

Control register for VRAM E

VRAM_F_CR

Control register for VRAM F

VRAM_G_CR

Control register for VRAM G

VRAM_H_CR

Control register for VRAM H

VRAM_I_CR

Control register for VRAM I