Module nds_sys::video

Video related functions and constants. Mostly “borrowed” from libnds’s video.h

Structs

DispCntFlags

Display control flags

Enums

VideoMode

Constants to be used for set_video_mode and set_video_mode_sub. The DS has 2 rendering engines, Main and Sub, that can be put in different modes (6 modes and 5 modes respectively). Mode6_2d and the ones suffixed “3D” are only valid for Main (The 3D engine renders on Background 0). Modes FB0-FB3 (“LCD” mode) map the Banks A-D respectively to pixels on screen.

Constants

BG_GFX

BG_GFX_SUB

BG_PALETTE

Background palette (Main)

BG_PALETTE_SUB

Background palette (Sub)

MAP_BASE_MASK

MAP_BASE_OFFSET

REG_DISPCNT

Display control register (main). Used to choose the mode for the Main Engine

REG_DISPCNT_SUB

Display control register (sub). Used to choose the mode for the Sub Engine

SCREEN_HEIGHT

SCREEN_WIDTH

TILES_BASE_MASK

TILES_BASE_OFFSET

VRAM_A

When VRAM A is in LCD mode, this points to the first pixel (top left most)

VRAM_A_CR

Control register for VRAM A

VRAM_B

When VRAM B is in LCD mode, this points to the first pixel (top left most)

VRAM_B_CR

Control register for VRAM B

VRAM_C

When VRAM C is in LCD mode, this points to the first pixel (top left most)

VRAM_C_CR

Control register for VRAM C

VRAM_D

When VRAM D is in LCD mode, this points to the first pixel (top left most)

VRAM_D_CR

Control register for VRAM D

VRAM_ENABLE

VRAM_E_CR

Control register for VRAM E

VRAM_F_CR

Control register for VRAM F

VRAM_G_CR

Control register for VRAM G

VRAM_H_CR

Control register for VRAM H

VRAM_I_CR

Control register for VRAM I

Functions

enable_main_background

set_video_mode⚠

Sets the video mode for the main engine.

set_video_mode_sub⚠

Sets the video mode for the sub engine. Mode 6 and 3D are invalid for the sub engine.

video_3d_enabled